Add manifest compression, optimizations
The manifest is now gzip compressed. Tests indicate the
compression is very effective since a lot of the items have
the same URL.

CHUNK_SIZE defaults to 1 now (down from 24). The python find()
method is poorly optimized. Since the manifest is compressed
the duplicate URL entries do not matter since they are squashed.
This also requires a smaller URL list.